Add files via upload
This tutorial is designed for beginner to intermediate tidyverse users. Many basic dplyr/tidyverse functions related to data management are demonstrated here; however, this is certainly not an all-inclusive list. Refer to the R documentation/vignettes for these packages for details about their features. This tutorial also briefly touches on some aspects of using ggplot (mostly within the context of plotting data from dplyr tibbles). The ggplot portion of this tutorial may be remedial for those familiar with the package. However, those without much ggplot experience may find it beneficial.